+++ kame/kame/pim6sd/debug.c
@@ -551,7 +551,8 @@ va_dcl
 	if (!debug)
 	    fprintf(fp, "%s: ", progname);
 	fprintf(fp, "%02d:%02d:%02d.%03ld %s", thyme->tm_hour,
-		thyme->tm_min, thyme->tm_sec, now.tv_usec / 1000, msg);
+		thyme->tm_min, thyme->tm_sec, (long int)now.tv_usec / 1000,
+		msg);
 	if (syserr == 0)
 	    fprintf(fp, "\n");
 	else
